♻️ Refactor test support and client internals into dedicated files
