+++ src/internal/curl_http_client.cpp
@@ -24,11 +24,10 @@
 
 #include "curl_http_client.hpp"
 
-#include "../../test/unit/curl_http_client_test_support.hpp"
+#include "curl_http_client_internal.hpp"
 #include "iqm_qdmi/constants.h"
 #include "logging.hpp"
 
-#include <chrono>
 #include <cstddef>
 #include <cstdint>
 #include <curl/curl.h>
@@ -36,115 +35,16 @@
 #include <nlohmann/json.hpp>
 #include <nlohmann/json_fwd.hpp>
 #include <string>
-#include <thread>
-#include <vector>
 
-namespace iqm {
+namespace iqm::internal {
 
-namespace {
-/**
- * @brief Logging policy used for non-success HTTP response handling.
- *
- * This enum controls whether response failures are surfaced at ERROR level for
- * required requests or downgraded to DEBUG for optional capability probes.
- */
-enum class ERROR_LOG_POLICY : uint8_t {
-  /// Log all errors at ERROR level (default for required requests)
-  LOG_AS_ERROR,
-  /// Log errors at DEBUG level (use for optional capability checks to avoid
-  /// alarming users with expected failures)
-  LOG_AS_DEBUG,
-};
-
-/// Number of retries to attempt after receiving HTTP 429.
-constexpr uint8_t RATE_LIMIT_RETRY_COUNT = 3;
-
-/// Delay in seconds between consecutive HTTP 429 retry attempts.
-constexpr uint8_t RATE_LIMIT_RETRY_DELAY_SECONDS = 2;
-
-/**
- * @brief Result of a single CURL request attempt.
- */
-struct Request_attempt_result {
-  /// The result returned by curl_easy_perform.
-  CURLcode curl_result = CURLE_OK;
-  /// The HTTP response code observed for the completed request.
-  int64_t response_code = 0;
-};
-
-/**
- * @brief Read the HTTP response code from a curl handle.
- *
- * @param curl The curl handle associated with the completed request.
- * @return The HTTP response code reported by libcurl.
- */
-int64_t Get_response_code(CURL *curl) {
-  int64_t response_code{};
-  curl_easy_getinfo(curl, CURLINFO_RESPONSE_CODE, &response_code);
-  return response_code;
-}
-
-/**
- * @brief Sleep for the given number of seconds before retrying a request.
- *
- * @param seconds The retry delay in seconds.
- */
-void Sleep_for_seconds(const uint8_t seconds) {
-  std::this_thread::sleep_for(std::chrono::seconds(seconds));
-}
-
-/**
- * @brief Function hooks used to override selected libcurl entry points.
- *
- * Tests use these hooks to force specific failure paths without requiring a
- * live network interaction.
- */
-struct Curl_api_hooks {
-  /// Hook for curl_easy_init.
-  CURL *(*easy_init)() = curl_easy_init;
-  /// Hook for curl_easy_perform.
-  CURLcode (*easy_perform)(CURL *) = curl_easy_perform;
-};
-
-// Helper for access to the mutable curl hook set
 Curl_api_hooks &Get_curl_api_hooks() {
   static Curl_api_hooks curl_api_hooks{};
   return curl_api_hooks;
 }
 
-// Helper to simulate curl_easy_init failure in tests
 CURL *Fail_curl_easy_init() { return nullptr; }
 
-// Helper for policy-aware error logging
-void Log_error(const ERROR_LOG_POLICY policy, const std::string &message) {
-  if (policy == ERROR_LOG_POLICY::LOG_AS_DEBUG) {
-    LOG_DEBUG(message);
-    return;
-  }
-  LOG_ERROR(message);
-}
-
-// Helper for CURL responses
-size_t Curl_write_callback(void *contents, const size_t size,
-                           const size_t nmemb, std::string *response) {
-  const size_t real_size = size * nmemb;
-  response->append(static_cast<char *>(contents), real_size);
-  return real_size;
-}
-
-// Helper for default HTTP request headers
-curl_slist *Default_headers(const std::string &bearer_token) {
-  curl_slist *headers = nullptr;
-  headers = curl_slist_append(headers, "User-Agent: IQM QDMI C++ API");
-  if (!bearer_token.empty()) {
-    // Set the "Authorization" header with the bearer token
-    headers =
-        curl_slist_append(headers, ("Authorization: " + bearer_token).c_str());
-  }
-  return headers;
-}
-
-// Helper for HTTP response code handling and logging
 int Handle_response_code(const int64_t response_code, const std::string &url,
                          const std::string &response,
                          const ERROR_LOG_POLICY error_log_policy) {
@@ -291,43 +191,34 @@ int Handle_response_code(const int64_t response_code, const std::string &url,
   return QDMI_ERROR_FATAL;
 }
 
-/**
- * @brief Execute a configured request and retry on HTTP 429 responses.
- *
- * The supplied callback performs one request attempt and returns both the
- * libcurl result and the resulting HTTP response code.
- */
-template <typename Perform_attempt, typename Sleep_function>
-int Perform_request_with_retries(const std::string &url, std::string &response,
-                                 const ERROR_LOG_POLICY error_log_policy,
-                                 Perform_attempt perform_attempt,
-                                 Sleep_function sleep_for_seconds) {
-  for (uint8_t attempt = 0; attempt <= RATE_LIMIT_RETRY_COUNT; ++attempt) {
-    response.clear();
-    const auto attempt_result = perform_attempt();
+} // namespace iqm::internal
 
-    if (attempt_result.curl_result != CURLE_OK) {
-      Log_error(error_log_policy, "curl_easy_perform() failed: " +
-                                      std::string(curl_easy_strerror(
-                                          attempt_result.curl_result)));
-      return QDMI_ERROR_FATAL;
-    }
+namespace {
 
-    if (attempt_result.response_code != 429 ||
-        attempt == RATE_LIMIT_RETRY_COUNT) {
-      return Handle_response_code(attempt_result.response_code, url, response,
-                                  error_log_policy);
-    }
+using iqm::internal::Attempt_result;
+using iqm::internal::ERROR_LOG_POLICY;
+using iqm::internal::Get_curl_api_hooks;
+using iqm::internal::Perform_request_with_retries;
+using iqm::internal::Request_attempt_result;
 
-    LOG_DEBUG("Request to URL '" + url +
-              "' hit HTTP 429 rate limiting; retrying in " +
-              std::to_string(RATE_LIMIT_RETRY_DELAY_SECONDS) +
-              " second(s) (attempt " + std::to_string(attempt + 1) + "/" +
-              std::to_string(RATE_LIMIT_RETRY_COUNT) + ")");
-    sleep_for_seconds(RATE_LIMIT_RETRY_DELAY_SECONDS);
+// Helper for CURL responses
+size_t Curl_write_callback(void *contents, const size_t size,
+                           const size_t nmemb, std::string *response) {
+  const size_t real_size = size * nmemb;
+  response->append(static_cast<char *>(contents), real_size);
+  return real_size;
+}
+
+// Helper for default HTTP request headers
+curl_slist *Default_headers(const std::string &bearer_token) {
+  curl_slist *headers = nullptr;
+  headers = curl_slist_append(headers, "User-Agent: IQM QDMI C++ API");
+  if (!bearer_token.empty()) {
+    // Set the "Authorization" header with the bearer token
+    headers =
+        curl_slist_append(headers, ("Authorization: " + bearer_token).c_str());
   }
-
-  return QDMI_ERROR_FATAL;
+  return headers;
 }
 
 int Perform_get_request(const std::string &url, const std::string &bearer_token,
@@ -350,33 +241,29 @@ int Perform_get_request(const std::string &url, const std::string &bearer_token,
   auto *headers = Default_headers(bearer_token);
   curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headers);
   const auto ret = Perform_request_with_retries(
-      url, response, error_log_policy,
-      [&]() -> Request_attempt_result {
-        const auto res = curl_api_hooks.easy_perform(curl);
-        if (res != CURLE_OK) {
-          return Request_attempt_result{.curl_result = res, .response_code = 0};
-        }
-        return Request_attempt_result{.curl_result = CURLE_OK,
-                                      .response_code = Get_response_code(curl)};
-      },
-      Sleep_for_seconds);
+      url, response, error_log_policy, [&]() -> Request_attempt_result {
+        return Attempt_result(curl, curl_api_hooks.easy_perform(curl));
+      });
   curl_easy_cleanup(curl); // NOLINT(misc-include-cleaner)
   curl_slist_free_all(headers);
   return ret;
 }
+
 } // namespace
+
+namespace iqm {
 
 int CurlHttpClient::get(const std::string &url, const std::string &bearer_token,
                         std::string &response) {
   return Perform_get_request(url, bearer_token, response,
-                             ERROR_LOG_POLICY::LOG_AS_ERROR);
+                             internal::ERROR_LOG_POLICY::LOG_AS_ERROR);
 }
 
 int CurlHttpClient::get_optional(const std::string &url,
                                  const std::string &bearer_token,
                                  std::string &response) {
   return Perform_get_request(url, bearer_token, response,
-                             ERROR_LOG_POLICY::LOG_AS_DEBUG);
+                             internal::ERROR_LOG_POLICY::LOG_AS_DEBUG);
 }
 
 /**
@@ -397,7 +284,7 @@ int CurlHttpClient::post(const std::string &url,
                          const std::string &data,
                          const std::string &extra_header) {
   LOG_INFO("Performing POST request to " + url);
-  auto &curl_api_hooks = Get_curl_api_hooks();
+  auto &curl_api_hooks = internal::Get_curl_api_hooks();
   CURL *curl = curl_api_hooks.easy_init();
   if (curl == nullptr) {
     LOG_ERROR("curl_easy_init() failed");
@@ -422,81 +309,15 @@ int CurlHttpClient::post(const std::string &url,
   } else {
     curl_easy_setopt(curl, CURLOPT_POSTFIELDS, "");
   }
-  const auto ret = Perform_request_with_retries(
-      url, response, ERROR_LOG_POLICY::LOG_AS_ERROR,
-      [&]() -> Request_attempt_result {
-        const auto res = curl_api_hooks.easy_perform(curl);
-        if (res != CURLE_OK) {
-          return Request_attempt_result{.curl_result = res, .response_code = 0};
-        }
-        return Request_attempt_result{.curl_result = CURLE_OK,
-                                      .response_code = Get_response_code(curl)};
-      },
-      Sleep_for_seconds);
+  const auto ret = internal::Perform_request_with_retries(
+      url, response, internal::ERROR_LOG_POLICY::LOG_AS_ERROR,
+      [&]() -> internal::Request_attempt_result {
+        return internal::Attempt_result(curl,
+                                        curl_api_hooks.easy_perform(curl));
+      });
   curl_easy_cleanup(curl); // NOLINT(misc-include-cleaner)
   curl_slist_free_all(headers);
   return ret;
 }
-
-namespace test_support {
-
-// NOLINTNEXTLINE(misc-use-internal-linkage)
-int Handle_response_code_for_testing(const int64_t response_code,
-                                     const std::string &url,
-                                     const std::string &response,
-                                     const bool use_debug_logging) {
-  return Handle_response_code(response_code, url, response,
-                              use_debug_logging
-                                  ? ERROR_LOG_POLICY::LOG_AS_DEBUG
-                                  : ERROR_LOG_POLICY::LOG_AS_ERROR);
-}
-
-// NOLINTNEXTLINE(misc-use-internal-linkage)
-void Set_curl_api_hooks_for_testing(CURL *(*easy_init)(),
-                                    CURLcode (*easy_perform)(CURL *)) {
-  auto &curl_api_hooks = Get_curl_api_hooks();
-  curl_api_hooks.easy_init = easy_init != nullptr ? easy_init : curl_easy_init;
-  curl_api_hooks.easy_perform =
-      easy_perform != nullptr ? easy_perform : curl_easy_perform;
-}
-
-// NOLINTNEXTLINE(misc-use-internal-linkage)
-void Enable_curl_easy_init_failure_for_testing() {
-  Set_curl_api_hooks_for_testing(Fail_curl_easy_init, nullptr);
-}
-
-// NOLINTNEXTLINE(misc-use-internal-linkage)
-Retry_test_result
-Retry_response_codes_for_testing(const std::vector<int64_t> &response_codes,
-                                 const std::string &url,
-                                 const bool use_debug_logging) {
-  size_t next_response_code_index = 0;
-  size_t sleep_call_count = 0;
-  std::string response;
-  const auto status_code = Perform_request_with_retries(
-      url, response,
-      use_debug_logging ? ERROR_LOG_POLICY::LOG_AS_DEBUG
-                        : ERROR_LOG_POLICY::LOG_AS_ERROR,
-      [&]() -> Request_attempt_result {
-        if (next_response_code_index >= response_codes.size()) {
-          return Request_attempt_result{.curl_result = CURLE_OK,
-                                        .response_code = 0};
-        }
-        const auto response_code = response_codes[next_response_code_index];
-        next_response_code_index++;
-        return Request_attempt_result{.curl_result = CURLE_OK,
-                                      .response_code = response_code};
-      },
-      [&]([[maybe_unused]] const uint8_t seconds) { sleep_call_count++; });
-  return Retry_test_result{.status_code = status_code,
-                           .sleep_call_count = sleep_call_count};
-}
-
-// NOLINTNEXTLINE(misc-use-internal-linkage)
-void Reset_curl_api_hooks_for_testing() {
-  Set_curl_api_hooks_for_testing(nullptr, nullptr);
-}
-
-} // namespace test_support
 
 } // namespace iqm
